Added ScaraLift control

+++ inc/hal/DynamixelMX12W.h
@@ -32,7 +32,7 @@ public:
     virtual void disableAndStop() override;
 
     void setRPM(float rpm);
-    void setSpeed(int16_t speed) override;
+    void setSpeed(int32_t speed) override;
 
     void stop() override;
 };
